requirements.txt
# Core
streamlit>=1.32.0
groq>=0.5.0
python-dotenv>=1.0.0

# RAG Pipeline (for Legal Q&A with document knowledge base)
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.3
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
PyPDF2>=3.0.0
pdfplumber>=0.10.0